refactor: Update OrderSummary component styling
This commit updates the styling of the OrderSummary component in OrderSummary.tsx. It adds the "xl:row-span-2" class to the parent div to adjust the layout for larger screens. It also adds the "xl:h-full" class to the inner div to ensure it takes up the full height of the parent container. These changes improve the visual presentation of the order summary section.